requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.12.0

# ── Experiment tracking ──────────────────────────────────────────────────────
mlflow>=2.10.0

# ── Serving ──────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# ── Config & serialization ───────────────────────────────────────────────────
pyyaml>=6.0
joblib>=1.3.0

# ── Visualization (notebooks only) ──────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Monitoring (Stage 5) ────────────────────────────────────────────────────
evidently>=0.4.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.0.0
httpx>=0.25.0
